requirements.txt
numpy==1.24.3
pycryptodome==3.18.0
matplotlib==3.7.1